+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-
+- [fix] Strings with backslash-escaped double quotes are
+  correctly converted to LUA code
 - [fix] Script generator escapes http header content
 - [feat] Skip `X-DevTools-Emulate-Network-Conditions-Client-Id` http header
 - [feat] After stop recording, open a tab to redirect the user into the app to save the script